move fetchwrapper to real common (used in viewer)
