+++ viewer/vueapp/src/components/connections/ConnectionsService.js
@@ -1,4 +1,4 @@
-import { cancelFetchWrapper } from '@/fetchWrapper.js';
+import { cancelFetchWrapper } from '@real_common/fetchWrapper.js';
 
 export default {
